Added role management, Hacky logout

+++ lib/pages/edit_event_page.dart
@@ -17,60 +17,41 @@ class EventsEditPage extends StatelessWidget {
       event: event,
       isNewEvent: false,
       onSaved: ({date, time, duration, title, venue, by, description, links}) {
-        showStatusDialog(context, "Event Updated!", () async {
-          await handler.updateEvent(
-            date: date,
-            time: time,
-            duration: duration,
-            title: title,
-            type: "",
-            venue: venue,
-            by: by,
-            description: description,
-            links: links,
-            documentID: event.documentID
-          );
-        });
+        showStatusDialog(
+          context: context, 
+          message: "Event Updated!", 
+          future: () async {
+            await handler.updateEvent(
+              date: date,
+              time: time,
+              duration: duration,
+              title: title,
+              type: "",
+              venue: venue,
+              by: by,
+              description: description,
+              links: links,
+              documentID: event.documentID
+            );
+          }
+        );
       },
       onDelete: () {
-        showDialog(
+        showConfirmDialog(
           context: context,
-          builder: (context) {
-            return SimpleDialog(
-              title: Text("Confirmation", style: TextStyle(fontFamily: "Poppings"),),
-              children: <Widget>[
-                ListTile(
-                  title: Text("Are you sure you want to delete the event?", style: TextStyle(fontFamily: "Poppins"),),                  
-                ),
-                SizedBox(height: 30,),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: <Widget>[
-                    FlatButton(
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-                      child: Text("Yes", style: TextStyle(fontFamily: "Poppins"),),
-                      onPressed: () {
-                        Navigator.of(context).pop();
-                        showStatusDialog(context, "Event Deleted!", () async {
-                          await handler.deleteEvent(event);
-                        });
-                      }
-                    ),
-                    FlatButton(
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-                      child: Text("No", style: TextStyle(fontFamily: "Poppins"),),
-                      onPressed: () {
-                        Navigator.of(context).pop();
-                      }
-                    )
-                  ],
-                )
-              ], 
+          message: "Are you sure you want to delete the event?",
+          onConfirm: () {
+            Navigator.of(context).pop();
+            showStatusDialog(
+              context: context, 
+              message: "Event Deleted!", 
+              future: () async {
+                await handler.deleteEvent(event);
+              }
             );
           }
         );
       },
     );
-
   }
 }
